Gets rid of old plot types
