--- core/src/main/java/org/cirdles/topsoil/plot/uth/evolution/EvolutionPlot.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2016 CIRDLES.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.cirdles.topsoil.plot.uth.evolution;
-
-import org.cirdles.commons.util.ResourceExtractor;
-import org.cirdles.topsoil.plot.JavaScriptPlot;
-
-/**
- * Created by johnzeringue on 12/1/15.
- */
-public class EvolutionPlot extends JavaScriptPlot {
-
-    private static final ResourceExtractor RESOURCE_EXTRACTOR
-            = new ResourceExtractor(EvolutionPlot.class);
-
-    private static final String RESOURCE_NAME = "EvolutionPlot.js";
-
-    public EvolutionPlot() {
-        super(
-                RESOURCE_EXTRACTOR.extractResourceAsPath(RESOURCE_NAME),
-                new EvolutionPlotDefaultProperties());
-    }
-
-}
